Pull test compiler out of FlutterPlatform
So that it can instead be used inside of `test/runner.dart`
